There are three changes in these files: 1. formatting; 2. Grammar builds; 3. Security fixes. For the formatting changes some files just got messed up and I was sick of looking at it. Grammar changes amount to simplifying the builds of the lex and y files. Now the sed calls in the Makefile have been removed - commented out to be completely removed later. 3. Nothing major on the security front but I made some changes all the same.

+++ Vector.cc
@@ -772,7 +772,8 @@ unsigned int Vector::buf2val(void **val)
     if (!val)
         throw InternalErr(__FILE__, __LINE__, "NULL pointer.");
 
-    int wid = width();          // This is the width computed using length(). The
+    unsigned int wid = static_cast<unsigned int>(width());
+    // This is the width computed using length(). The
     // length() property is changed when a projection
     // constraint is applied. Thus this is the number of
     // bytes in the buffer given the current constraint.
